pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dino"
version = "0.1.0"
description = "Chrome Dino Game RL Agent"
requires-python = ">=3.14"
dependencies = [
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "mss>=9.0.0",
    "stable-baselines3>=2.0.0",
    "pynput>=1.7.0",
    "PyGObject>=3.42.0",
    "pillow>=10.0.0",
    "dbus-python>=1.3.0",
    "opencv-python-headless>=4.11.0.86",
    "evdev>=1.9.2",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM", "RUF",]

[tool.setuptools]
py-modules = ["main", "pipewire_capture"]

[tool.mypy]
disable_error_code = ["import-untyped"]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pygobject-stubs>=2.12.0",
    "ruff>=0.14.11",
    "types-pynput>=1.7.7.20241221",
]
